+++ phtree/phtree_multimap_d_test_filter.cc
@@ -168,10 +168,18 @@ struct FilterCount {
         ++f_destruct_;
     }
 
-    [[nodiscard]] constexpr bool IsEntryValid(const PhPoint<DIM>&, const T& value) {
-        last_known = const_cast<T&>(value);
+    template <typename BucketT>
+    [[nodiscard]] constexpr bool IsEntryValid(const PhPoint<DIM>&, const BucketT& bucket) {
+        assert(!bucket.empty());
         return true;
     }
+
+    template <typename T2>
+    [[nodiscard]] bool IsBucketEntryValid(const PhPoint<DIM>&, const T2& value) {
+        last_known = value;
+        return true;
+    }
+
     [[nodiscard]] constexpr bool IsNodeValid(const PhPoint<DIM>&, int) {
         return true;
     }
@@ -253,7 +261,11 @@ struct CallbackCount {
 
 template <dimension_t DIM, typename T>
 struct FilterConst {
-    [[nodiscard]] constexpr bool IsEntryValid(const PhPoint<DIM>&, const T& value) const {
+    template <typename BucketT>
+    [[nodiscard]] constexpr bool IsEntryValid(const PhPoint<DIM>&, const BucketT&) const {
+        return true;
+    }
+    [[nodiscard]] constexpr bool IsBucketEntryValid(const PhPoint<DIM>&, const T& value) {
         assert(value._i == 1);
         return true;
     }
@@ -276,6 +288,18 @@ struct CallbackConst {
               << std::endl;
 }
 
+/*
+ * General comment: We are testing several thing here.
+ * - If we pass lvalue filters/callbacks/... we want to ensure that they do not get copied or
+ *   moved at all. We need to ensure that the lvalue argument is the same instance that is
+ *   used internally by the iterator.
+ * - If we pass a rvalue filters/callbacks/..., preventing copies/moves is harder. We are testing
+ *   somewhat arbitrarily for a limit of 3 moves/copies per argument.
+ * - We want to ensure that both rvalue/lvalue arguments work.
+ * - We also do some limited testing that it works with 'const' trees.
+ * - Finally, we test separately that the old legacy filters still work
+ */
+
 TEST(PhTreeTest, TestFilterAPI_FOR_EACH) {
     // Test edge case: only one entry in tree
     PhPointD<3> p{1, 2, 3};
@@ -287,6 +311,7 @@ TEST(PhTreeTest, TestFilterAPI_FOR_EACH) {
     // rvalue
     tree.for_each(callback, filter);
     ASSERT_EQ(static_id, 1);
+    ASSERT_EQ(filter.last_known._i, 1);
     ASSERT_EQ(2, f_construct_ + f_default_construct_);
     ASSERT_EQ(0, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
     f_reset_id_counters();
@@ -295,7 +320,7 @@ TEST(PhTreeTest, TestFilterAPI_FOR_EACH) {
     tree.for_each(CallbackCount<3>(), FilterCount<3, Id>());
     ASSERT_EQ(static_id, 1);
     ASSERT_EQ(2, f_construct_ + f_default_construct_);
-    ASSERT_LE(1, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
+    ASSERT_GE(4, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
     f_reset_id_counters();
 
     // const Tree: just test that it compiles
@@ -321,6 +346,7 @@ TEST(PhTreeTest, TestFilterAPI_FOR_EACH_WQ) {
     // lvalue
     tree.for_each(qb, callback, filter);
     ASSERT_EQ(static_id, 1);
+    ASSERT_EQ(filter.last_known._i, 1);
     ASSERT_EQ(2, f_construct_ + f_default_construct_);
     ASSERT_EQ(0, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
     f_reset_id_counters();
@@ -329,7 +355,7 @@ TEST(PhTreeTest, TestFilterAPI_FOR_EACH_WQ) {
     tree.for_each({{1, 2, 3}, {4, 5, 6}}, CallbackCount<3>{}, FilterCount<3, Id>());
     ASSERT_EQ(static_id, 1);
     ASSERT_EQ(2, f_construct_ + f_default_construct_);
-    ASSERT_LE(1, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
+    ASSERT_GE(4, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
     f_reset_id_counters();
 
     // const Tree: just test that it compiles
@@ -351,14 +377,15 @@ TEST(PhTreeTest, TestFilterAPI_BEGIN) {
     FilterCount<3, Id> filter{};
     // lvalue
     ASSERT_EQ(tree.begin(filter)->_i, 1);
+    ASSERT_EQ(filter.last_known._i, 1);
     ASSERT_EQ(1, f_construct_ + f_default_construct_);
-    ASSERT_EQ(0, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
+    ASSERT_GE(0, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
     f_reset_id_counters();
 
     // rvalue
     ASSERT_EQ(tree.begin(FilterCount<3, Id>())->_i, 1);
     ASSERT_EQ(1, f_construct_ + f_default_construct_);
-    ASSERT_LE(1, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
+    ASSERT_GE(2, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
     f_reset_id_counters();
 
     // const Tree: just test that it compiles
@@ -381,6 +408,7 @@ TEST(PhTreeTest, TestFilterAPI_WQ) {
     FilterCount<3, Id> filter{};
     // lvalue
     ASSERT_EQ(tree.begin_query(qb, filter)->_i, 1);
+    ASSERT_EQ(filter.last_known._i, 1);
     ASSERT_EQ(1, f_construct_ + f_default_construct_);
     ASSERT_EQ(0, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
     f_reset_id_counters();
@@ -388,7 +416,7 @@ TEST(PhTreeTest, TestFilterAPI_WQ) {
     // rvalue
     ASSERT_EQ(tree.begin_query({{1, 2, 3}, {4, 5, 6}}, FilterCount<3, Id>())->_i, 1);
     ASSERT_EQ(1, f_construct_ + f_default_construct_);
-    ASSERT_LE(1, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
+    ASSERT_GE(2, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
     f_reset_id_counters();
 
     // const Tree: just test that it compiles
@@ -411,6 +439,7 @@ TEST(PhTreeTest, TestFilterAPI_KNN) {
     DistanceCount<3> dist_fn{};
     // lvalue
     ASSERT_EQ(tree.begin_knn_query(3, {2, 3, 4}, dist_fn, filter)->_i, 1);
+    ASSERT_EQ(filter.last_known._i, 1);
     ASSERT_EQ(2, f_construct_ + f_default_construct_);
     ASSERT_EQ(0, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
     f_reset_id_counters();
@@ -418,14 +447,14 @@ TEST(PhTreeTest, TestFilterAPI_KNN) {
     // rvalue
     ASSERT_EQ(tree.begin_knn_query(3, {2, 3, 4}, DistanceCount<3>{}, FilterCount<3, Id>())->_i, 1);
     ASSERT_EQ(2, f_construct_ + f_default_construct_);
-    ASSERT_LE(0, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
+    ASSERT_GE(2 * 3, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
     f_reset_id_counters();
 
     // rvalue #2
     auto a = tree.begin_knn_query<DistanceCount<3>, FilterCount<3, Id>>(3, {2, 3, 4})->_i;
     ASSERT_EQ(a, 1);
     ASSERT_EQ(2, f_construct_ + f_default_construct_);
-    ASSERT_LE(0, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
+    ASSERT_GE(2 * 3, f_copy_construct_ + f_move_construct_ + f_copy_assign_ + f_move_assign_);
     f_reset_id_counters();
 
     // const Tree: just test that it compiles
@@ -436,4 +465,127 @@ TEST(PhTreeTest, TestFilterAPI_KNN) {
     // rvalue
     ASSERT_EQ(treeC.begin_knn_query(3, {2, 3, 4}, DistanceCount<3>{}, FilterConst<3, Id>())->_i, 1);
     f_reset_id_counters();
+}
+
+template <dimension_t DIM>
+double distance(const TestPoint<DIM>& p1, const TestPoint<DIM>& p2) {
+    double sum2 = 0;
+    for (dimension_t i = 0; i < DIM; ++i) {
+        double d2 = p1[i] - p2[i];
+        sum2 += d2 * d2;
+    }
+    return sqrt(sum2);
+};
+
+template <dimension_t DIM>
+void referenceSphereQuery(
+    std::vector<TestPoint<DIM>>& points,
+    TestPoint<DIM>& center,
+    double radius,
+    std::set<size_t>& result) {
+    for (size_t i = 0; i < points.size(); i++) {
+        auto& p = points[i];
+        if (distance(center, p) <= radius) {
+            result.insert(i);
+        }
+    }
+}
+
+// We use 'int&' because gtest does not compile with assertions in non-void functions.
+template <dimension_t DIM>
+void testSphereQuery(TestPoint<DIM>& center, double radius, size_t N, int& result) {
+    TestTree<DIM, size_t> tree;
+    std::vector<TestPoint<DIM>> points;
+    populate(tree, points, N);
+
+    std::set<size_t> referenceResult;
+    referenceSphereQuery(points, center, radius, referenceResult);
+
+    result = 0;
+    auto filter = FilterMultiMapSphere(center, radius, tree.converter());
+    for (auto it = tree.begin(filter); it != tree.end(); it++) {
+        auto& x = *it;
+        ASSERT_GE(x, 0);
+        ASSERT_EQ(referenceResult.count(x), 1);
+        result++;
+    }
+    ASSERT_EQ(referenceResult.size(), result);
+}
+
+TEST(PhTreeMMDFilterTest, TestSphereQuery0) {
+    const dimension_t dim = 3;
+    TestPoint<dim> p{-10000, -10000, -10000};
+    int n = 0;
+    testSphereQuery<dim>(p, 0.1, 100, n);
+    ASSERT_EQ(0, n);
+}
+
+TEST(PhTreeMMDFilterTest, TestSphereQueryMany) {
+    const dimension_t dim = 3;
+    TestPoint<dim> p{0, 0, 0};
+    int n = 0;
+    testSphereQuery<dim>(p, 1000, 1000, n);
+    ASSERT_GT(n, 400);
+    ASSERT_LT(n, 800);
+}
+
+TEST(PhTreeMMDFilterTest, TestSphereQueryAll) {
+    const dimension_t dim = 3;
+    TestPoint<dim> p{0, 0, 0};
+    int n = 0;
+    testSphereQuery<dim>(p, 10000, 1000, n);
+    ASSERT_EQ(1000, n);
+}
+
+
+// We use 'int&' because gtest does not compile with assertions in non-void functions.
+template <dimension_t DIM>
+void testSphereQueryForEach(TestPoint<DIM>& center, double radius, size_t N, int& result) {
+    TestTree<DIM, size_t> tree;
+    std::vector<TestPoint<DIM>> points;
+    populate(tree, points, N);
+
+    std::set<size_t> referenceResult;
+    referenceSphereQuery(points, center, radius, referenceResult);
+
+    result = 0;
+    auto filter = FilterMultiMapSphere(center, radius, tree.converter());
+    auto callback = [&result, &referenceResult](PhPointD<DIM>, const size_t& x) {
+        ASSERT_GE(x, 0);
+        ASSERT_EQ(referenceResult.count(x), 1);
+        ++result;
+    };
+    tree.for_each(callback, filter);
+//    for (auto it = tree.begin(filter); it != tree.end(); it++) {
+//        auto& x = *it;
+//        ASSERT_GE(x, 0);
+//        ASSERT_EQ(referenceResult.count(x), 1);
+//        result++;
+//    }
+    ASSERT_EQ(referenceResult.size(), result);
+}
+
+TEST(PhTreeMMDFilterTest, TestSphereQueryForEach0) {
+    const dimension_t dim = 3;
+    TestPoint<dim> p{-10000, -10000, -10000};
+    int n = 0;
+    testSphereQueryForEach<dim>(p, 0.1, 100, n);
+    ASSERT_EQ(0, n);
+}
+
+TEST(PhTreeMMDFilterTest, TestSphereQueryForEachMany) {
+    const dimension_t dim = 3;
+    TestPoint<dim> p{0, 0, 0};
+    int n = 0;
+    testSphereQueryForEach<dim>(p, 1000, 1000, n);
+    ASSERT_GT(n, 400);
+    ASSERT_LT(n, 800);
+}
+
+TEST(PhTreeMMDFilterTest, TestSphereQueryForEachAll) {
+    const dimension_t dim = 3;
+    TestPoint<dim> p{0, 0, 0};
+    int n = 0;
+    testSphereQueryForEach<dim>(p, 10000, 1000, n);
+    ASSERT_EQ(1000, n);
 }